Updated README.md with many details
This is a first pass at some of the documentation, specifically the main README.md file. It adds links to other documentation that we'll be adding elsewhere.

We are making these updates to make sure that we are all on the same page as we begin work on the project.