Add Production MySQL Credential to Application Properties
We discussed as a group and have decided to include hardcoded credentials for our MySQL database. Normally we wouldn't do this since there are security risks in doing so. We decided that in order to gain a path of least resistance towards having a functioning application, we've decided to ignore industry best practices and check these credentials into version control. Later in the semester, if time allows, we will come back around and work on obfuscating and securing the database and application better.

* Fixed a missing constructor that is needed for JPA to work properly