Expose All Actuator Endpoints
Actuator is a built in tool that comes with Spring Boot that allows you to monitor your application. It exposes configuration, metrics, information, health info, etc. etc. 

I am exposing this on a different port than 8080 so we can potentially lock it down so you can't hit it from the outside world.