Added some useful backend authentication links
While working on the first pass at backend authorization, I came across several pages which helped me as well as those which could potentially help us in the future as we move towards using Spring Security, etc.

I didn't want to lose these links so I'm "bookmarked" them here in the repository so they can be accessed in the future.