Finished Implementing Walmart API Refresh
The scheduled task has been implemented. On the specified schedule it will select all of the items from the database and stream over them updating the values.

We make the assumption that whatever the most current api request says is the truth that it _is_ the truth and just overwrite _all_ values with the current api's values. It didn't seem worth the logic, at least not at this point, to check each individual field and look for changes so we just make the assumption and overwrite everything.

Some additional adjustments:
* Disabled the very verbose apache logging. I left the line commented out as it can be helpful for troubleshooting if we need it again in the future.
* Set a few log statements to Trace since we don't need them all the time.
* Inlined a few of the inserts so there wasn't redundant, unneeded objects being created.