Also return Buyer/SellerItem Information from new endpoints
Based on feedback from the frontend, we also need to make sure and _still_ include the BuyerItem / SellerItem information (such as the desired price, etc) when we respond with the new DetailedItems in context of the Buyer/Seller endpoints.

Note that, just like the previous commit, there is some question about what to do to handle items which don't have both pieces (both an Item _and_ a WalmartItem). I've added the same TODO in the code there to come back in the future and handle it.