Add Tests for ItemController
This adds a set of tests for the `ItemController`.

The pattern of testing this controller should be followed for other Controller tests. 

It uses `MockMvc` to issue and verify the requests, `@SpringBootTest` which launches the app context, and Spring Boot's `SpringRunner` test runner which wraps JUnit and adds additional logging, etc.

These tests are run with the rest of the unit tests during the maven build.