Split WalmartService into Service/Client & Added Tests
While writing tests for the existing `WalmartService` I found that the testing was awkward which suggested an issue with architecture.

I split the WalmartService into a Service (interacts with the repository, hold the business logic, etc) and a Client which actually makes remote calls to Walmart's API. This simplified testing and made the code more isolated and easier to maintain in the long run.

I also added many tests for the service and the client.