Added initial Dockerfile and Docker Image build instructions
This is an initial Dockerfile for creating an image out of the backend app. There are probably optimizations that can be made to it so that the resulting file is smaller (right now it is about 145mb). It also isn't clear whether we'd be alright using OpenJDK instead of the Oracle-provided "official" JDK, or possibly even a JRE instead of a JDK.

This also adds some initial build instructions for building the docker image from the Dockerfile. These are also a first-go and can likely be improved as well.