Conditionally build Docker Image, default to skip building of image
Because we ran into several configuration issues on several Windows machines, we've decided to make the building of a docker image as part of the maven build conditional. By default we will NOT build an image. Configuration was added so that it only occurs if a flag of `-Ddocker.skip=false` is included in the maven install command.