Update Backend to 0.3.0 and add more detail in backend Maven docks
This is a pretty good place to version up to 0.3.0 for the backend api.

Not only do we need to be on Maven 3.5 but we also need to make sure that developers know how to change IntelliJ to use that later version installed externally. I've added docs which go over this.