Use an internal H2 db (instead of prod db) for Backend tests
Connecting to our Production Database for unit and component tests wasn't _ever_ a great idea, but it worked. Until we started doing build using Travis-CI...

Its still unclear _why_ it is happening with Travis-CI but when it runs the tests as it is bringing up the Spring Application Context it initializes a connection pool with the MySQL RDS db. But for some reason it establishes a TON of connections. We don't see this on localhost while testing against the Prod DB, nor does it happen every time on the Travis builds...

A better way to test at this level is actually just with an embedded H2 database. This way it still wires everything up that it needs but just creates its own internal db so we aren't reliant on the Prod DB at all. This _should_ fix our problems with our builds, and is actually just better practice to follow anyway.