Upgrade Spring Boot to 2.1.3
Upgrading to the latest release of Spring Boot. There was a vulnerability in `2.0.5` which has started to be exploited. Upgrading to `2.0.8` mitigates the issue, but we may as well go all the way to the latest release of `2.1.3`. Tested with build and a quick smattering of manual tests and it appears to work without any issues.